parser/cgen: default struct field values

+++ vlib/strconv/ftoa/f32_str.v
@@ -22,14 +22,14 @@ module ftoa
 // dec32 is a floating decimal type representing m * 10^e.
 struct Dec32 {
 mut:
-	m u32 = u32(0)
+	m u32 = 0
 	e int = 0
 }
 
 // support union for convert f32 to u32
 union Uf32 {
 mut:
-	f f32 = f32(0)
+	f f32 = 0
 	u u32
 }
